+++ lib/typescript/commonjs/core/deriveDbKey.d.ts
@@ -1,0 +1,3 @@
+import type { CollectionModel } from '../types';
+export declare const deriveDbKey: (model: CollectionModel<any, any>, scope?: object) => readonly unknown[];
+//# sourceMappingURL=deriveDbKey.d.ts.map